blog/requirements.txt
beautifulsoup4==4.6.1
bs4==0.0.1
Django==2.0.7
django-ckeditor==5.6.1
django-jinja==2.4.1
django-js-asset==1.1.0
django-rest-framework==0.1.0
django-taggit==0.22.2
djangorestframework==3.8.2
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.2.0
pytz==2018.5
six==1.11.0
Unidecode==1.0.22
